requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.0
pandas==2.1.0
numpy==1.24.3
xgboost==2.0.0
scikit-learn==1.3.0
pymongo==4.5.0
python-dotenv==1.0.0
holidays==0.38
huggingface-hub==0.20.2
